+++ bdtsim/protocol/delgado/__init__.py
@@ -14,21 +14,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from .protocol import Protocol, DEFAULT_ASSET_PRICE
-from .protocol_manager import ProtocolManager
-from .exceptions import ProtocolError, ProtocolInitializationError, ProtocolExecutionError
-from .delgado import Delgado
-from .fairswap import FairSwap, FairSwapReusable
-from .simplepayment import SimplePayment, SimplePaymentPrepaid, SimplePaymentPostpaid
-from .smartjudge import SmartJudge
+from .delgado import Delgado, DelgadoLibrary, DelgadoReusable
 
 __all__ = [
-    'Protocol', 'DEFAULT_ASSET_PRICE',
-    'ProtocolManager',
-    'ProtocolError', 'ProtocolInitializationError', 'ProtocolExecutionError',
-    'Delgado', 'DelgadoReusable',
-    'FairSwap', 'FairSwapReusable',
-    'SimplePayment', 'SimplePaymentPrepaid', 'SimplePaymentPostpaid',
-    'SmartJudge'
+    'Delgado',
+    'DelgadoReusable',
+    'DelgadoLibrary'
 ]